requirements.txt
streamlit
pandas
numpy
prophet
crewai
crewai-tools
langchain-google-genai
matplotlib
openpyxl
python-dotenv